refactor(calendar): extract push-conflict-retry from sync-service (M6 T7)
`sync-service.ts` crossed the 800-line eslint cap after M6 T6's
sync-health additions. Extracted three pure functions to a new
sibling file `push-conflict-retry.ts`:

- `loadSourceAsGoogleEvent(db, target)` — maps a CalendarSyncTarget
  to a `GoogleCalendarUpsertEventInput`
- `pushEventWithConflictRetry(db, target, client, calendarId, binding)`
  — the M3 412 conflict loop (3 retries, merges remote, marks
  binding `remoteVersion='conflict'` on exhaustion)
- `mergeRemoteEventIntoLocal(db, eventId, remote)` — the M3 field-
  clock-aware merge that keeps the local doc clock untouched

No behaviour change. sync-service drops from 949→760 lines; the
extracted module is ~180 lines and self-contained (no circular
imports — it depends on mappers, field-merge, and change-events but
not on sync-service). Pre-existing test surface continues to cover
both files.
